refactor: move watch endpoint to its own route file, decoupling from CU routes
