+++ assistant/src/runtime/routes/computer-use-routes.ts
@@ -1,8 +1,7 @@
 /**
  * HTTP route handlers for computer use session lifecycle.
  *
- * These endpoints expose CU session management and watch
- * observation functionality over HTTP.
+ * These endpoints expose CU session management over HTTP.
  *
  * All CU write operations require the `chat.write` scope.
  */
@@ -37,18 +36,6 @@ export interface ComputerUseDeps {
   sharedRequestTimestamps: number[];
   /** Sequence tracker for CU observations (per session). */
   cuObservationParseSequence: Map<string, number>;
-  /** Handle a watch observation. */
-  handleWatchObservation: (params: {
-    watchId: string;
-    sessionId: string;
-    ocrText: string;
-    appName?: string;
-    windowTitle?: string;
-    bundleIdentifier?: string;
-    timestamp: number;
-    captureIndex: number;
-    totalExpected: number;
-  }) => Promise<void>;
 }
 
 // ---------------------------------------------------------------------------
@@ -373,71 +360,6 @@ async function handleTaskSubmit(
   );
 }
 
-/**
- * POST /v1/computer-use/watch — send a watch observation.
- *
- * Body: { watchId, sessionId, ocrText, appName?, windowTitle?,
- *         bundleIdentifier?, timestamp, captureIndex, totalExpected }
- */
-async function handleWatchObservationRoute(
-  req: Request,
-  deps: ComputerUseDeps,
-): Promise<Response> {
-  const body = (await req.json()) as {
-    watchId?: string;
-    sessionId?: string;
-    ocrText?: string;
-    appName?: string;
-    windowTitle?: string;
-    bundleIdentifier?: string;
-    timestamp?: number;
-    captureIndex?: number;
-    totalExpected?: number;
-  };
-
-  if (!body.watchId || typeof body.watchId !== "string") {
-    return httpError("BAD_REQUEST", "watchId is required", 400);
-  }
-  if (!body.sessionId || typeof body.sessionId !== "string") {
-    return httpError("BAD_REQUEST", "sessionId is required", 400);
-  }
-  if (!body.ocrText || typeof body.ocrText !== "string") {
-    return httpError("BAD_REQUEST", "ocrText is required", 400);
-  }
-  if (typeof body.timestamp !== "number") {
-    return httpError("BAD_REQUEST", "timestamp is required", 400);
-  }
-  if (typeof body.captureIndex !== "number") {
-    return httpError("BAD_REQUEST", "captureIndex is required", 400);
-  }
-  if (typeof body.totalExpected !== "number") {
-    return httpError("BAD_REQUEST", "totalExpected is required", 400);
-  }
-
-  try {
-    await deps.handleWatchObservation({
-      watchId: body.watchId,
-      sessionId: body.sessionId,
-      ocrText: body.ocrText,
-      appName: body.appName,
-      windowTitle: body.windowTitle,
-      bundleIdentifier: body.bundleIdentifier,
-      timestamp: body.timestamp,
-      captureIndex: body.captureIndex,
-      totalExpected: body.totalExpected,
-    });
-
-    return Response.json({ ok: true });
-  } catch (err) {
-    const message = err instanceof Error ? err.message : String(err);
-    log.error(
-      { err, watchId: body.watchId },
-      "Failed to handle watch observation via HTTP",
-    );
-    return httpError("INTERNAL_ERROR", message, 500);
-  }
-}
-
 // ---------------------------------------------------------------------------
 // Route definitions
 // ---------------------------------------------------------------------------
@@ -476,12 +398,6 @@ export function computerUseRouteDefinitions(deps: {
       method: "POST",
       policyKey: "computer-use/tasks",
       handler: async ({ req }) => handleTaskSubmit(req, getDeps()),
-    },
-    {
-      endpoint: "computer-use/watch",
-      method: "POST",
-      policyKey: "computer-use/watch",
-      handler: async ({ req }) => handleWatchObservationRoute(req, getDeps()),
     },
   ];
 }
